requirements.txt
pandas==1.3.4
Pillow==9.1.1
streamlit==1.8.1
